cmd/llvm-obfuscator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.0.post1
typer==0.9.0
rich==13.7.0
pyyaml==6.0.1
pydantic>=2.7,<3
jinja2==3.1.3
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
requests==2.31.0
